+++ src/components/NewLabel.js
@@ -7,14 +7,20 @@ export default function NewLabel(props) {
     title,
     desc,
     thumbnailUrl,
+    className,
   } = props;
   return (
-    <div className="w-auto inline-block p-2">
-      <img className="w-full rounded-md" src={thumbnailUrl} alt="Thumbnail" />
+    <div className={`${className} p-2`}>
+      {thumbnailUrl
+        && <img className="w-full rounded-md" src={thumbnailUrl} alt="Thumbnail" />}
       <div>
-        <a href={`/new/${path}`}>
+        {thumbnailUrl ? (
+          <a href={`/new/${path}`}>
+            <h1 className="font-bold text-xl">{title}</h1>
+          </a>
+        ) : (
           <h1 className="font-bold text-xl">{title}</h1>
-        </a>
+        )}
         <p className="text-lg font-light">{desc}</p>
       </div>
     </div>
@@ -22,8 +28,15 @@ export default function NewLabel(props) {
 }
 
 NewLabel.propTypes = {
-  path: PropTypes.string.isRequired,
+  path: PropTypes.string,
   title: PropTypes.string.isRequired,
   desc: PropTypes.string.isRequired,
-  thumbnailUrl: PropTypes.string.isRequired,
+  thumbnailUrl: PropTypes.string,
+  className: PropTypes.string,
+};
+
+NewLabel.defaultProps = {
+  className: 'w-auto inline-block',
+  thumbnailUrl: null,
+  path: null,
 };